Stabilization/MW: document converting from MultiWii to our units
The MultiWii control loop has lots of arbitrary constants. Here I work
out how to convert from one representation to the other so people can
reuse those settings. In our code we use physically we defined units
such as accounting for the loop rate and the units of the gyro.